:dizzy: move app.get('/') to routes/home.js
