+++ lib/db.js
@@ -66,6 +66,33 @@ const self = {
 		});
 	},
 
+	curr_person: (req) => {
+		return new Promise(async (resolve, reject) => {
+
+			try {
+
+				if ('session' in req &&
+				    'person' in req.session) {
+
+					let query = await self.query(`
+						SELECT *
+						FROM person
+						WHERE id = $1
+					`, [req.session.person.id]);
+
+					if (query.rows.length > 0) {
+						return resolve(query.rows[0]);
+					}
+				}
+
+				resolve(null);
+
+			} catch(err) {
+				reject(err);
+			}
+		});
+	},
+
 	get_context: (id_or_slug) => {
 		return new Promise(async (resolve, reject) => {
 
@@ -98,6 +125,256 @@ const self = {
 				}
 
 				resolve(query.rows[0]);
+
+			} catch(err) {
+				reject(err);
+			}
+		});
+	},
+
+	get_contexts: (person, message_id) => {
+		return new Promise(async (resolve, reject) => {
+
+			try {
+
+				let query;
+				const contexts = {};
+
+				if (person) {
+
+					query = await self.query(`
+						SELECT context.*
+						FROM member, context
+						WHERE member.active = true
+						  AND member.person_id = $1
+						  AND member.context_id = context.id
+					`, [person.id]);
+
+					contexts.member_of = query.rows;
+
+					let subgroups = {};
+					for (let context of contexts.member_of) {
+						if (context.parent_id) {
+							if (! subgroups[context.parent_id]) {
+								subgroups[context.parent_id] = [];
+							}
+							subgroups[context.parent_id].push(context);
+						}
+					}
+
+					for (let context of contexts.member_of) {
+						if (subgroups[context.id]) {
+							context.subgroups = subgroups[context.id];
+						} else {
+							context.subgroups = [];
+						}
+					}
+
+					let thread = null;
+					let context_id = null;
+
+					if (person.context_id) {
+						context_id = person.context_id;
+					}
+
+					if (message_id) {
+						query = await self.query(`
+							SELECT message.*,
+								   person.slug AS person_slug, person.name AS person_name
+							FROM message, person
+							WHERE (message.id = $1 OR in_reply_to = $1)
+							  AND message.person_id = person.id
+							ORDER BY created
+						`, [message_id]);
+						thread = query.rows;
+						context_id = thread[0].context_id;
+					}
+
+					if (context_id) {
+
+						query = await self.query(`
+							SELECT *
+							FROM context
+							WHERE id = $1
+						`, [context_id]);
+
+						let current = query.rows[0];
+
+						if (thread) {
+							current.thread = thread;
+						}
+
+						contexts.current = await self.add_context_details(current);
+					}
+				}
+
+				resolve(contexts);
+
+			} catch(err) {
+				reject(err);
+			}
+		});
+	},
+
+	add_context_details: (context, before_id) => {
+		return new Promise(async (resolve, reject) => {
+
+			try {
+
+				let query;
+
+				if (! context) {
+					return resolve(context);
+				}
+
+				if (context.thread) {
+
+					await self.add_message_details(context.thread);
+					context.messages = context.thread.splice(0, 1);
+					context.messages[0].replies = context.thread;
+
+				} else {
+
+					let before_clause = '';
+					let values = [context.id];
+					if (before_id) {
+						before_clause = 'AND message.id < $2';
+						values.push(before_id);
+					}
+
+					query = await self.query(`
+						SELECT message.*,
+						       person.name AS person_name, person.slug AS person_slug
+						FROM message, person
+						WHERE message.context_id = $1
+						  ${before_clause}
+						  AND message.person_id = person.id
+						  AND message.in_reply_to IS NULL
+						ORDER BY message.created DESC
+						LIMIT 10
+					`, values);
+
+					context.messages = query.rows;
+
+					query = await self.query(`
+						SELECT COUNT(id) AS total_messages
+						FROM message
+						WHERE context_id = $1
+						  AND in_reply_to IS NULL
+					`, [context.id]);
+
+					context.total_messages = query.rows[0].total_messages;
+
+					await self.add_message_details(context.messages);
+				}
+
+				query = await self.query(`
+					SELECT member.person_id, person.name, person.slug
+					FROM member, person
+					WHERE member.active = true
+					  AND member.context_id = $1
+					  AND member.person_id = person.id
+					ORDER BY member.updated DESC
+				`, [context.id]);
+
+				context.members = query.rows;
+
+				query = await self.query(`
+					SELECT *
+					FROM context
+					WHERE parent_id = $1
+				`, [context.id]);
+				context.subgroups = query.rows;
+
+				if (context.parent_id) {
+					context.parent = await self.get_context(context.parent_id);
+				}
+
+				resolve(context);
+
+			} catch(err) {
+				reject(err);
+			}
+		});
+	},
+
+	add_message_details: (messages) => {
+		return new Promise(async (resolve, reject) => {
+
+			try {
+
+				if (! 'length' in messages || messages.length == 0) {
+					return resolve(messages);
+				}
+
+				let ids = messages.map(msg => msg.id);
+				let placeholders = [];
+				for (let i = 0; i < ids.length; i++) {
+					placeholders.push('$' + (i + 1));
+				}
+
+				placeholders = placeholders.join(', ');
+
+				let query = await self.query(`
+					SELECT in_reply_to AS id,
+						   COUNT(id) AS reply_count
+					FROM message
+					WHERE in_reply_to IN (${placeholders})
+					GROUP BY in_reply_to
+				`, ids);
+
+				let replies = {};
+				for (let reply of query.rows) {
+					replies[reply.id] = reply.reply_count;
+				}
+
+				query = await self.query(`
+					SELECT target_id, content, created
+					FROM facet
+					WHERE target_id IN (${placeholders})
+					  AND target_type = 'message'
+					  AND facet_type = 'revision'
+					ORDER BY created DESC
+				`, ids);
+
+				let revisions = {};
+				for (let revision of query.rows) {
+
+					revision.created = new Date(revision.created).toISOString();
+
+					if (! revisions[revision.target_id]) {
+						revisions[revision.target_id] = [];
+					}
+					revisions[revision.target_id].push({
+						created: revision.created,
+						content: revision.content
+					});
+				}
+
+				for (let message of messages) {
+
+					message.created = new Date(message.created).toISOString();
+					message.updated = new Date(message.updated).toISOString();
+
+					if (message.id in replies) {
+						message.reply_count = replies[message.id];
+					} else {
+						message.reply_count = 0;
+					}
+
+					if (revisions[message.id]) {
+						message.revisions = revisions[message.id];
+					} else {
+						message.revisions = [];
+					}
+					message.revisions.unshift({
+						created: message.updated,
+						content: message.content
+					});
+					message.revision_dates = message.revisions.map(rev => rev.created);
+				}
+
+				resolve(messages);
 
 			} catch(err) {
 				reject(err);
